Made dashboard.py as multipage entry file, added current_ride dashboard to pages folder
